refactor(ui): extract message-list formatting seam
